// client-id/src/app/core/models/thuoc.ts
export class Thuoc {
    proId: string;
    proName: string;
    proNo: string;
    proProducer: string;
    proPack: string;
    proUnit: string;
    proPrice: number;
    typeId: number;
    typeName: string;

    constructor() {
        this.proId = null;
        this.proName = null;
        this.proNo = null;
        this.proProducer = null;
        this.proPack = null;
        this.proUnit = null;
        this.proPrice = 0;
        this.typeId = null;
        this.typeName = null;
    }
}