// client-id/src/app/core/models/nhap-hang.ts
export class NhapHang {
    impId: string;
    impDate: Date;
    impNote: string;
    userId: string;
    supId: string;
    supName: string;
    impTotal: number;
    proId: string;
    proName: string;
    quantity: number;

    constructor() {
        this.impId = null;
        this.impDate = null;
        this.impNote = null;
        this.userId = null;
        this.supId = null;
        this.supName = null;
        this.impTotal = null;
        this.proId = null;
        this.proName = null;
        this.quantity = null;
    }
}